+++ newPupilScreen.pas
@@ -1,0 +1,30 @@
+unit newPupilScreen;
+
+interface
+
+uses
+  Winapi.Windows, Winapi.Messages, System.SysUtils, System.Variants, System.Classes, Vcl.Graphics,
+  Vcl.Controls, Vcl.Forms, Vcl.Dialogs, Vcl.StdCtrls;
+
+type
+  TForm19 = class(TForm)
+    firstnameEdit: TEdit;
+    lastnameEdit: TEdit;
+    Label1: TLabel;
+    Label2: TLabel;
+    create: TButton;
+    Label3: TLabel;
+  private
+    { Private declarations }
+  public
+    { Public declarations }
+  end;
+
+var
+  Form19: TForm19;
+
+implementation
+
+{$R *.dfm}
+
+end.