+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# 1.0.6
+* auth logic added
+* users logic added 
+* classes creating
+* new user creating
+
 # 1.0.5
 * dates save and read for each subject
 * new date validator